move classes into specific packages
