+++ src/main/java/br/com/financialtoolapi/application/usecases/business/expense/CreateOrUpdateExpenseUseCase.java
@@ -8,8 +8,8 @@ import br.com.financialtoolapi.application.dtos.in.ExpenseInputDto;
 import br.com.financialtoolapi.application.exceptions.ResourceNotFoundException;
 import br.com.financialtoolapi.application.exceptions.ValidationDataException;
 import br.com.financialtoolapi.application.mapper.ExpenseMapper;
-import br.com.financialtoolapi.application.usecases.business.FindExpenseCategoryByIdUseCase;
-import br.com.financialtoolapi.application.usecases.business.FindUserByAccountIdUseCase;
+import br.com.financialtoolapi.application.usecases.business.expenseCategory.FindExpenseCategoryByIdUseCase;
+import br.com.financialtoolapi.application.usecases.business.userAccount.FindUserByAccountIdUseCase;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
